Cleanup and modernization of source code
Massive commit, sorry about that. And there's another one coming.

I am refactoring pretty much everything using modern Javascript,
because it's a cool project and it deserves it. This work is not
finished, but at this point the code works without errors so it's
a good checkpoint.

Also three important changes:

 * Removing massive code duplication in `background.js`, by
   importing other files with `importScripts`.

 * Removing old code for migrating configuration. To whoever is
   using Bookkeeper from 2017, and today, in 2025, out of the blue
   feels like updating to V4, I apologize: you will have to track
   all your buildings again ¯\_(ツ)_/¯

 * Initializing a sane configuration on first installation.